requirements.txt
# requirements.txt
# Service IA - Détection d'Objets Perdus

# === FRAMEWORK WEB ===
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
websockets==12.0

# === MACHINE LEARNING ===
torch==2.1.0
torchvision==0.16.0
numpy==1.24.3

# === VISION PAR ORDINATEUR ===
opencv-python==4.8.1.78
Pillow==10.0.1

# === VALIDATION ET SCHEMAS ===
pydantic==2.4.2
pydantic-settings==2.0.3

# === ASYNC ET CONCURRENCE ===
asyncio==3.4.3
aiofiles==23.2.1

# === UTILITAIRES ===
python-dotenv==1.0.0
pathlib==1.0.1
uuid==1.30
logging==0.4.9.6

# === MONITORING ET PERFORMANCE ===
psutil==5.9.6
GPUtil==1.4.0

# === TRAITEMENT DE DONNEES ===
pandas==2.1.1
scipy==1.11.3

# === TESTS ===
pytest==7.4.2
pytest-asyncio==0.21.1
httpx==0.25.0

# === DÉVELOPPEMENT ===
black==23.9.1
flake8==6.1.0
mypy==1.6.1

# === OPTIONNEL - BASES DE DONNÉES ===
# sqlalchemy==2.0.21
# alembic==1.12.0
# psycopg2-binary==2.9.7  # PostgreSQL
# redis==5.0.0  # Cache

# === OPTIONNEL - MONITORING AVANCÉ ===
# prometheus-client==0.17.1
# elastic-apm==6.18.1

# === OPTIONNEL - DÉPLOIEMENT ===
# gunicorn==21.2.0
# docker==6.1.3

# === OPTIONNEL - SÉCURITÉ ===
# python-jose[cryptography]==3.3.0
# passlib[bcrypt]==1.7.4

# === NOTES D'INSTALLATION ===
# 
# Installation de base:
# pip install -r requirements.txt
#
# Installation avec GPU (CUDA):
# pip install torch torchvision --index-url https://download.pytorch.org/whl/cu118
# 
# Installation pour développement:
# pip install -r requirements.txt
# pip install pytest pytest-asyncio httpx black flake8 mypy
#
# Installation production:
# pip install -r requirements.txt
# pip install gunicorn prometheus-client
#
# Vérification GPU:
# python -c "import torch; print(f'CUDA available: {torch.cuda.is_available()}')"